Alteração da estrutura e construtor da classe Database

+++ Documents/NetBeansProjects/PerceptronSingleLayerLogicalE/src/Model/DataBase.java
@@ -30,7 +30,7 @@ public class DataBase {
         }
     }
 
-    public double[][] getData() {
+    public int[][] getData() {
         return data;
     }
